security refactor
